+++ update_service/lib/src/update_service.dart
@@ -1,6 +1,7 @@
 import 'dart:developer';
 import 'dart:io' as io;
 
+import 'package:update_service/src/app_version.dart';
 import 'package:url_launcher/url_launcher.dart';
 
 import '../update_service.dart';
@@ -33,13 +34,26 @@ class UpdateService {
     );
   }
 
-  Future<UpdateStatus> checkStatus({bool? skipUpdate}) async {
+  /// Checks the update status of the application.
+  ///
+  /// - If `skipUpdate` is set to `true`, the version check will be bypassed,
+  ///   and the status will be returned as `upToDate`, regardless of the actual version.
+  ///
+  /// - If `forceUpdateVersion` is specified, any app version lower than this
+  ///   will require a forced update. This parameter expects a version format like "1.0.2".
+  ///
+  /// Returns an [UpdateStatus] indicating the update status of the app, which could
+  /// be `required`, `optional`, `upToDate`, or `unknown` in case of an error.
+  Future<UpdateStatus> checkStatus({
+    bool? skipUpdate,
+    String? forceUpdateVersion,
+  }) async {
     if (skipUpdate == true) {
       return UpdateStatus.upToDate();
     }
 
     try {
-      return await _checkStatus();
+      return await _checkStatus(forceUpdateVersion);
     } catch (ex, s) {
       log('[update_service] Failed to check update status',
           error: ex, stackTrace: s);
@@ -47,7 +61,7 @@ class UpdateService {
     }
   }
 
-  Future<UpdateStatus> _checkStatus() async {
+  Future<UpdateStatus> _checkStatus(String? forceUpdateVersion) async {
     final localVersion = await _packageInfoService.getAppVersion();
     final packageName = await _packageInfoService.getPackageName();
     if (packageName == null) {
@@ -72,24 +86,35 @@ class UpdateService {
       return UpdateStatus.unknown();
     }
 
-    return _computeUpdateStatus(localVersion, publishedVersion);
+    return _computeUpdateStatus(
+      localVersion,
+      publishedVersion,
+      forceUpdateVersion,
+    );
   }
 
   UpdateStatus _computeUpdateStatus(
-      String localVersion, String publishedVersion) {
-    final [pubMajor, pubMinor, pubPatch] =
-        publishedVersion.split('.').map(int.parse).toList();
-    final [localMajor, localMinor, localPatch] =
-        localVersion.split('.').map(int.parse).toList();
+    String localVersion,
+    String publishedVersion,
+    String? forceUpdateVersion,
+  ) {
+    final local = AppVersion.parse(localVersion);
+    final published = AppVersion.parse(publishedVersion);
+    final force = _getForceUpdateVersion(forceUpdateVersion);
 
-    if (pubMajor > localMajor) {
+    if (force != null && local < force && published >= force) {
       return UpdateStatus.required(localVersion, publishedVersion);
-    } else if (pubMinor > localMinor ||
-        (pubPatch > localPatch && pubMinor == localMinor)) {
+    }
+
+    if (local < published) {
       return UpdateStatus.optional(localVersion, publishedVersion);
-    } else {
+    }
+
+    if (local >= published) {
       return UpdateStatus.upToDate();
     }
+
+    return UpdateStatus.unknown();
   }
 
   void openStore() async {
@@ -128,4 +153,19 @@ class UpdateService {
 
   Future<bool> _launchPlayStoreWebUri(String packageName) => launchUrl(
       Uri.parse('https://play.google.com/store/apps/details?id=$packageName'));
+
+  AppVersion? _getForceUpdateVersion(String? forceUpdateVersion) {
+    if (forceUpdateVersion == null) return null;
+    try {
+      return AppVersion.parse(forceUpdateVersion);
+    } catch (ex, s) {
+      log(
+        'Invalid forceUpdateVersion format',
+        error: ex,
+        stackTrace: s,
+        name: 'UpdateService',
+      );
+      return null;
+    }
+  }
 }